Restore Scintilla header files
Restore Sci_Position.h & ScintillaStructures.h (ref: https://github.com/notepad-plus-plus/notepad-plus-plus/pull/12327#issuecomment-1296020276)
And add R styles.

+++ scintilla/include/Sci_Position.h
@@ -18,7 +18,7 @@ typedef ptrdiff_t Sci_Position;
 typedef size_t Sci_PositionU;
 
 // For Sci_CharacterRange  which is defined as long to be compatible with Win32 CHARRANGE
-typedef intptr_t Sci_PositionCR;
+typedef long Sci_PositionCR;
 
 #ifdef _WIN32
 	#define SCI_METHOD __stdcall
